+++ src/run.ts
@@ -197,6 +197,8 @@ export class Run {
       testImage = "node:20-slim"
     } else if ("jmeter" in this.state.config.test) {
       testImage = this.state.config.test.jmeter.image ?? "justb4/jmeter:latest"
+    } else if ("cucumber" in this.state.config.test) {
+      testImage = this.state.config.test.cucumber.image ?? "testplatform/cucumber-runner:latest"
     } else {
       testImage = this.state.config.test.image
     }